+++ __tests__/01_basic_spec.js
@@ -1,7 +1,7 @@
 import React, { Suspense } from 'react';
 import { render, cleanup } from '@testing-library/react';
 
-import { prefetch, useFetch } from '../src/index';
+import { createFetch, useFetch } from '../src/index';
 
 const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
 
@@ -9,13 +9,14 @@ describe('basic spec', () => {
   afterEach(cleanup);
 
   it('should have a function', () => {
-    expect(prefetch).toBeDefined();
+    expect(createFetch).toBeDefined();
     expect(useFetch).toBeDefined();
   });
 
   it('should create a component', async () => {
     fetch.mockResponse(JSON.stringify('test data'));
-    const result = prefetch('http://...');
+    const fetchUrl = async url => (await fetch(url)).json();
+    const result = createFetch(fetchUrl, 'http://...');
     const DisplayData = () => {
       const { data } = useFetch(result);
       return <span>Data: {data}</span>;