+++ src/index.d.ts
@@ -4,15 +4,25 @@ type Fallback = ReactNode | ((error?: Error, retry?: () => void) => ReactNode);
 
 export class ErrorBoundary extends Component<{ fallback: Fallback }> {}
 
-type Input<Data> = string | Request | (() => Promise<Data>);
-type Result<Data> = { data: Data };
-
-export const prefetch: <Data>(input: Input<Data>) => Result<Data>;
-
-export const initialize: <Data>(initialData: Data) => Result<Data>;
-
-export const useFetch: <Data>(
-  initialResult: Result<Data>
-) => Result<Data> & {
-  refetch: (input: Input<Data>) => void;
+type Result<Data, Input> = {
+  data: Data;
+  refetch: (input: Input) => Result<Data, Input>;
 };
+
+type CreateFetch = {
+  <Data, Input>(
+    fetchFunc: (input: Input) => Promise<Data>,
+    initialInput: Input
+  ): Result<Data, Input>;
+  <Data, Input>(
+    fetchFunc: (input: Input) => Promise<Data>,
+    initialInput: null,
+    initialData: Data
+  ): Result<Data, Input>;
+};
+
+export const createFetch: CreateFetch;
+
+export const useFetch: <Data, Input>(
+  initialResult: Result<Data, Input>
+) => Result<Data, Input>;